// generated/src/app/api-models/subscriber-update-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SubscriberUpdateSetup {
  seqSubscId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  status: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}